Changed types for `fromEvent`, add mapSync to AsyncStream
Hopefully the type change allows for some "more" type safety around
event stream messages.  Unfortunately, the `@types/node` types are not
particularly helpful, and give that this was the use-case, that was kind
of frustrating.  But this should hopefully not particularly affect
browser events... I think.

Also added `mapSync` to async streams.  I used it for an identity
mapping, to do type assertions, but I think this probably isn't
necessary.  The function call overhead is almost certainly annoying.